Split up Core into Rx, Var, Ctx and Obs
