+++ scalarx/shared/src/main/scala/rx/Rx.scala
@@ -2,9 +2,8 @@ package rx
 
 import rx.opmacros.Factories
 
-import scala.annotation.compileTimeOnly
 import scala.collection.mutable
-import scala.util.{Success, Try}
+import scala.util.Try
 
 /**
   * A reactive value of type [[T]]. Keeps track of triggers and
@@ -218,146 +217,4 @@ object Rx {
     override def toString = s"Rx@${Integer.toHexString(hashCode()).take(2)}($now)"
   }
 
-}
-
-trait Var[T] extends Rx[T] {
-  def update(newValue: T): Unit
-
-  def update(f: T => T):Unit = {
-    update(f(now))
-  }
-
-  private[rx] var value: T
-
-  private[rx] def depth: Int = 0
-
-  def toTry: Try[T] = util.Success(now)
-
-  private[rx] override def recalc(): Unit = propagate()
-
-  private[rx] override def kill(): Unit = {
-    clearDownstream()
-  }
-
-  override def toString = s"Var@${Integer.toHexString(hashCode()).take(2)}($now)"
-}
-
-object Var {
-  /**
-    * Create a [[Var]] from an initial value
-    */
-  def apply[T](initialValue: T): Var[T] = new Base(initialValue)
-
-  /**
-    * Set the value of multiple [[Var]]s at the same time; in doing so,
-    * reduces the redundant updates that would normally occur setting
-    * them one by one
-    */
-  def set(args: Assignment[_]*): Unit = {
-    args.foreach(_.set())
-    Rx.doRecalc(
-      args.flatMap(_.v.downStream),
-      args.flatMap(_.v.observers)
-    )
-  }
-
-  case class Assignment[T](v: Var[T], value: T) {
-    def set(): Unit = {
-      v.value = value
-    }
-  }
-
-  /**
-    * Encapsulates the act of setting of a [[Var]] to a value, without
-    * actually setting it.
-    */
-  object Assignment {
-    implicit def tuple2Assignment[T](t: (Var[T], T)): Assignment[T] = {
-      Assignment(t._1, t._2)
-    }
-
-    implicit def tuples2Assignments[T](ts: Seq[(Var[T], T)]): Seq[Assignment[T]] = {
-      ts.map(t => Assignment(t._1, t._2))
-    }
-  }
-
-  /**
-    * A smart variable that can be set manually, and will notify downstream
-    * [[Rx]]s and run any triggers whenever its value changes.
-    */
-  class Base[T](initialValue: T) extends Var[T] {
-
-    private[rx] var value = initialValue
-
-    override def now: T = value
-
-    /**
-      * Sets the value of this [[Var]] and runs any triggers/notifies
-      * any downstream [[Rx]]s to update
-      */
-    def update(newValue: T): Unit = {
-      if (value != newValue) {
-        value = newValue
-        Rx.doRecalc(downStream, observers)
-      }
-    }
-  }
-}
-
-object Ctx{
-
-  object Data extends Generic[Data]{
-    @compileTimeOnly("No implicit Ctx.Data is available here!")
-    implicit object CompileTime extends Data(???)
-  }
-  class Data(rx0: => Rx.Dynamic[_]) extends Ctx(rx0)
-
-  object Owner extends Generic[Owner]{
-    @compileTimeOnly("No implicit Ctx.Owner is available here!")
-    object CompileTime extends Owner(???)
-
-    object Unsafe extends Owner(???){
-      implicit val Unsafe: Ctx.Owner.Unsafe.type = this
-    }
-    /**
-      * Dark magic. End result is the implicit ctx will be one of
-      *  1) The enclosing RxCtx, if it exists
-      *  2) RxCtx.Unsafe, if in a "static context"
-      *  3) RxCtx.CompileTime, if in a "dynamic context" (other macros will rewrite CompileTime away)
-      */
-    @compileTimeOnly("@}}>---: A rose by any other name.")
-    implicit def voodoo: Owner = macro Factories.automaticOwnerContext[rx.Ctx.Owner]
-  }
-
-  class Owner(rx0: => Rx.Dynamic[_]) extends Ctx(rx0)
-
-  class Generic[T] {
-    def safe(): T = macro Factories.buildSafeCtx[T]
-  }
-}
-
-/**
-  * An implicit scope representing a "currently evaluating" [[Rx]]. Used to keep
-  * track of dependencies or ownership.
-  */
-class Ctx(rx0: => Rx.Dynamic[_]){
-  lazy val contextualRx: Rx.Dynamic[_] = rx0
-}
-
-
-/**
-  * Wraps a simple callback, created by `trigger`, that fires when that
-  * [[Rx]] changes.
-  */
-class Obs(val thunk: () => Unit, upstream: Rx[_]) {
-
-  var dead = false
-
-  /**
-    * Stop this observer from triggering and allow it to be garbage-collected
-    */
-  def kill(): Unit = {
-    upstream.observers.remove(this)
-    dead = true
-  }
 }
